OMG IT ACTUALLY DRAWS ALL THE THUNGS

+++ src/Model/Coastlines/Coastline.java
@@ -22,7 +22,7 @@ public class Coastline extends OSMWay {
     @Override
     public Path2D toPath2D() {
         List<Point2D> points = new ArrayList<>();
-        float longFactor = CoastlineFactory.getLongitudeFactor();
+        double longFactor = CoastlineFactory.getLongitudeFactor();
         for (Point2D node: this) {
             points.add(new Point2D.Double(node.getX()*longFactor,node.getY()));
         }
